video: change is_editing to datetime
For some browser, like safari, page cache will prevent to trigger the
onbeforeunload event.

Use datetime instead of a boolean for is_editing. So if is_editing is nil or
Time.now > is_editing, we allow the editing page to be accessed. And the
expire time is 2 hours.

A little tricky.